feat: optimize chat colors for readability
I removed the .75 opacity as this way of adjusting colors has the drawback of hurting readability when using a background.

The italics color must be as visually distinct as possible from the regular color, while still being readable. I have picked #a3a3a3, which is the very darkest color which still has an AAA contrast ratio against the background color #171717. It is the objective best choice. It wasn't an existing CSS variable, but going outside the palette is worth it here.