feat: "Anonymize" and "Screenshot" chat features
This commit adds an "Anonymize chat" toggle to the chat settings which
hides avatars and replaces all occurrences of the user's handle with
"User 1". In multi-user room, each user has a different number so
it is still clear who is speaking.

This commit also adds a "Screenshot loaded msgs" link in chat settings,
which opens a new tab with a screenshot of every loaded message.