feat: screenshot loaded msgs
This adds a "Screenshot loaded msgs" button in chat settings.
Unlike the previous implementation this doesn't directly modify the DOM.
Since rendering a large screenshot can take a while, this also changes
the "Screenshot loaded msgs" to an italic loading message while the
screenshot is generating. Clicking the button while the screenshot is
generating does nothing.