fix: memory book searches in reverse when logged in
+ tweak: separate memory book entries with newline instead of '. '

Explanation:

If you were logged in, memory book search would start from the beginning
of the conversation history instead of the end.

This was caused by the browser prompt-building to be subtly different from the
server prompt-building because of the function `getLinesForPrompt`
returning lines "newest first", unlike the rest of prompt related code.

This would then be overridden by the generate endpoint handler for
logged-in users.

What this PR does:
- Make browser prompt-building use the same line order as server
  prompt-building
- Fix memory search for both cases by removing a reverse() inside
  buildMemoryPrompt

I've tested that I didn't break chatting for guests or logged-in users
(screenshots in PR), but you might want to verify this and any other
area you suspect I might have broken.