feat: support character book in chat
If the character has a character book, then we make the memory prompt with the character book first. then whatever is left of the token budget for memory book, we then use that to make the memory prompt with the world book. Character book is inserted *after* world book, since text located lower in the prompt has a higher weight on current LLMs.