refactor: change the way character book is selected and edited
Previously, setting a Character Book to be an existing Agnai Memory Book
would cause confusion if updating the Agnai Memory Book: the character
book would not update until you re-selected the Agnai Memory Book. This
commit changes the UX to make it clear that when you select an Agnai Memory
Book, a copy is created. This commit also adds a way to edit the bundled
character book, using EditMemoryForm in a modal.