fix: wide imgs breaking ui with avatar wrap off
When avatar wrap is off, fix bug where wide images would cause the avatar column to go up to 10000px height.

This is due to the avatar float causing the message text element's width to contain the avatar column, so when images reach 100% of their container, this 100% does not exclude the avatar-column width. We can force the message text element to be adjacent to the avatar column with `overflow-hidden` (for some reason), but we only include this when wrap around is off, or it obviously cancels the wraparound.