feat: {{#lowpriority}}
Adds a new {{#lowpriority}}{{/lowpriority}} syntax for text that is only
inserted if there still is token budget remaining for it after inserting
conversation history.

Suggested usage:
{{#if example_dialogue}}
{{#lowpriority}}
This is how {{char}} speaks:
{{example_dialogue}}
{{/lowpriority}}
{{/if}}

Added documentation for it in Placeholder Definitions, as well as
documentation for {{#insert}}. This adds buttons for insert and
lowpriority in the Prompt Template string builder.

NOT SUPPORTED ON CLAUDE/OPENAI CHAT.